==> 0_data_summary_and_generate_dioDataCSVs.ipynb <==
## Data Summary

from collections import defaultdict
from collections import Counter
import numpy as np

intents = defaultdict(int)
attr_dict = defaultdict(Counter)

with open('./Multiwoz/WOZ_train_ans.txt') as f:
    for line in f:
        line = line.strip()
        comps = line.split('|')
        # count intent
        intent = comps[0]
        intents[intent] += 1
        
        if len(comps) > 1:
            attrs = comps[1:]
        
            for attr in attrs:
                attr_type, ans = tuple(attr.split('='))
                attr_dict[attr_type].update([ans])

intents

attr_dict.keys()

attr_dict['hotel-area']

attr_dict['hotel-internet']

attr_dict['hotel-parking']

attr_dict['hotel-pricerange']

attr_dict['hotel-type']

attr_dict['hotel-stars']

attr_dict['restaurant-area']

attr_dict['restaurant-pricerange']

attr_dict['restaurant-area']

## Convert txt files into dataframe

hotel_attrs = [
    "hotel-area",
    "hotel-internet",
    "hotel-parking",
    "hotel-name",
    "hotel-pricerange",
    "hotel-type",
    "hotel-stars"
]

restaurant_attrs = [
    'restaurant-food', 
    'restaurant-pricerange', 
    'restaurant-area', 
    'restaurant-name'
]

HEADERS = ['find_hotel', 'find_restaurant', 'utts'] + hotel_attrs + restaurant_attrs

import csv

def write_data_to_df(file_in_utt, file_in_ans, file_out):
    """Write utterances and answers into DataFrame format based on 
    the given file_in_utt and file_in_ans, and name the output file file_out
    """
    with open(file_out, 'w') as fout:
        writer = csv.DictWriter(fout, fieldnames=HEADERS)
        writer.writeheader()

        with open(file_in_ans) as f_ans, open(file_in_utt) as f_utt:
            for ans, utt in zip(f_ans, f_utt):
                row = {attr: None for attr in HEADERS}
                utt = utt.strip()
                row['utts'] = utt

                comps = ans.split('|')
                intent = comps[0].strip('\n')
                row[intent] = 1

                if len(comps) > 1:
                    for comp in comps[1:]:
                        attr_type, ans = tuple(comp.split('='))
                        row[attr_type] = ans.strip('\n')

                writer.writerow(row)

def write_test_df(file_in_utt, file_out):
    """Write the test utterances in file_in_utt to a DataFrame, file_out"""
    with open(file_out, 'w') as fout:
        writer = csv.DictWriter(fout, fieldnames=HEADERS)
        writer.writeheader()

        with open(file_in_utt) as f_utt:
            for utt in f_utt:
                row = {attr: None for attr in HEADERS}
                utt = utt.strip()
                row['utts'] = utt
                writer.writerow(row)

# write_data_to_df('../Multiwoz/WOZ_train_utt.txt', '../Multiwoz/WOZ_train_ans.txt', '../data/dioData_train.csv')

# write_data_to_df('../Multiwoz/WOZ_dev_utt.txt', '../Multiwoz/WOZ_dev_ans.txt', '../data/dioData_dev.csv')

# write_test_df('../Multiwoz/WOZ_test_utt.txt', '../data/dioData_test.csv')
