==> 1_intent_classification.ipynb <==
## Intent Classification: Simple ML approach

## Get Started

from sklearn.pipeline import Pipeline, make_pipeline
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.svm import SVC
from sklearn.dummy import DummyClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score
from sklearn.ensemble import RandomForestClassifier
import pandas as pd
import numpy as np

## Load data

multiwoz_path = "../Multiwoz/"

def load_data(file_name):
  '''Given file name, read the file and return a list of lines'''
  with open(multiwoz_path+file_name, "r", encoding="utf-8") as f:
    data = f.read().splitlines()

  return data

X_train = load_data("WOZ_train_utt.txt")
y_train = load_data("WOZ_train_ans.txt")
X_dev = load_data("WOZ_dev_utt.txt")
y_dev = load_data("WOZ_dev_ans.txt")
X_test = load_data("WOZ_test_utt.txt")

X_train[:5]

y_train[:5]

## Construct dataset for binary classification

def prepare_for_binary_classification(X, y):
  '''Split tags in y, pick only the first tag and remove `find_`'''
  y_binary = [answer.split("|")[0].split("_")[-1] for answer in y]

  return X, y_binary

X_binary_train, y_binary_train = prepare_for_binary_classification(X_train, y_train)
X_binary_dev, y_binary_dev = prepare_for_binary_classification(X_dev, y_dev)
X_binary_test = X_test

y_binary_train[:5]

## Baseline Score

def fit_transform_and_predict(X_train, y_train, X_dev, vec, cls):
  '''
  Generate predictions based on the arguments.
  X_train (list) : train data
  y_train (list) : train tags
  X_dev (list) : development (or test) data to be predicted
  vec : vectorizer
  cls : classifier
  '''
  # vectorize
  X_train_enc = vec.fit_transform(X_train, y_train)
  X_dev_enc = vec.transform(X_dev)

  # train
  cls.fit(X_train_enc, y_train)

  # predict
  preds = cls.predict(X_dev_enc)

  return preds

cls = DummyClassifier(strategy="most_frequent")
vec = CountVectorizer()

dev_pred = fit_transform_and_predict(X_binary_train, y_binary_train, X_binary_dev, vec, cls)

print("dev acc (Dummy): ", accuracy_score(y_binary_dev, dev_pred))

## Try different ML systems: SVC, Logistic Regression, and Random Forest

# svc
svc = SVC()
dev_svc_pred = fit_transform_and_predict(X_binary_train, y_binary_train, X_binary_dev, vec, svc)
print("dev acc (SVC): ", accuracy_score(y_binary_dev, dev_svc_pred))

# lr
vec = CountVectorizer(ngram_range = (1, 1), 
                                        lowercase = True,
                                        stop_words = 'english')
lr = LogisticRegression()
dev_lr_pred = fit_transform_and_predict(X_binary_train, y_binary_train, X_binary_dev, vec, lr)
print("dev acc (lr): ", accuracy_score(y_binary_dev, dev_lr_pred))

# random forest

vec = CountVectorizer(ngram_range = (1, 2), 
                                        lowercase = True, 
                                        stop_words = 'english')
rf = RandomForestClassifier()
dev_rf_pred = fit_transform_and_predict(X_binary_train, y_binary_train, X_binary_dev, vec, rf)
print("dev acc (rf): ", accuracy_score(y_binary_dev, dev_lr_pred))

We chose `random forest classifier` for this task.

## Generate binary intent predictions for the development set and the test set

train_rf_pred = fit_transform_and_predict(X_binary_train, y_binary_train, X_binary_train, vec, rf)
dev_rf_pred = fit_transform_and_predict(X_binary_train, y_binary_train, X_binary_dev, vec, rf)
test_rf_pred = fit_transform_and_predict(X_binary_train, y_binary_train, X_binary_test, vec, rf)

assert len(dev_rf_pred) == 413
assert len(test_rf_pred) == 400
print("Success!")

with open("../predictions/0_train_binary_intent_pred.txt", "w", encoding="utf-8") as fout:
  for pred in train_rf_pred:
    fout.write(pred+"\n")

with open("../predictions/1_dev_binary_intent_pred.txt", "w", encoding="utf-8") as fout:
  for pred in dev_rf_pred:
    fout.write(pred+"\n")

with open("../predictions/2_test_binary_intent_pred.txt", "w", encoding="utf-8") as fout:
  for pred in test_rf_pred:
    fout.write(pred+"\n")

## Inspect weights of each feature

pipe_lr = make_pipeline(
    CountVectorizer(ngram_range = (1, 1), 
                                        lowercase = True,
                                        stop_words = 'english'),
    LogisticRegression()
)

pipe_lr.fit(X_binary_train, y_binary_train)
pipe_lr.score(X_binary_dev, y_binary_dev)

weights = pipe_lr.named_steps['logisticregression'].coef_.flatten()
vocab = pipe_lr.named_steps['countvectorizer'].get_feature_names()
data={'features':vocab[-500:-480], 'coefficient':weights[-500:-480]}
inds = np.argsort(pipe_lr.named_steps['logisticregression'].coef_.flatten())

pipe_lr.classes_

hotal_words = [vocab[index] for index in inds[:20]]
restaurants_words = [vocab[index] for index in inds[-20:]]
hotel_words_weights = [(weights[index]) for index in inds[:20]]
restaurants_words_weights = [(weights[index]) for index in inds[-20:]]
pd.DataFrame(
    {
        "hotel feats": hotal_words, "hotel weights": hotel_words_weights,
        "restaurants feats": restaurants_words,"restaurants weights": restaurants_words_weights,
    })
