==> 4_ml-based-approach.ipynb <==
## ML-approach to hotel-type, restaraunt-pricerange, and hotel-pricerange

## 1. Hotel-type predictions

import pandas as pd
import numpy as np
from sklearn.pipeline import Pipeline, make_pipeline
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.svm import SVC
from sklearn.dummy import DummyClassifier
from sklearn.metrics import accuracy_score

def read_df_with_pred_intent(df_filename, intent_filename):
    '''combine intent predictions with the given dataframe'''
    df = pd.read_csv(df_filename)
    with open(intent_filename) as f:
        binary = f.read().splitlines()
    df['intent'] = binary
    return df

train_df = read_df_with_pred_intent('dioData_train.csv', 'train_binary_intent_pred.txt')
dev_df = read_df_with_pred_intent('dioData_dev.csv', 'dev_binary_intent_pred.txt')

hotel_train_df = train_df[train_df['intent'] == 'hotel']
hotel_dev_df = dev_df[dev_df['intent'] == 'hotel']

X_train = hotel_train_df['utts']
y_train = hotel_train_df['hotel-type'].replace(np.nan, 'dontknow')
X_dev = hotel_dev_df['utts']
y_dev = hotel_dev_df['hotel-type'].replace(np.nan, 'dontknow')

### Baseline DummyClassifier

pipe = make_pipeline(CountVectorizer(), DummyClassifier())

pipe.fit(X_train, y_train)

pipe.score(X_train, y_train)

### SVC

pipe_svc = make_pipeline(CountVectorizer(), SVC())
pipe_svc.fit(X_train, y_train)
print("training score: ", pipe_svc.score(X_train, y_train))
print("validation score: ", pipe_svc.score(X_dev, y_dev))

### Random Forest

# Random Forest
from sklearn.ensemble import RandomForestClassifier
pipe_rf = make_pipeline(CountVectorizer(ngram_range = (1, 3), 
                                        lowercase = True, 
                                        stop_words = 'english'
#                                         max_features = 2500
                                       ), 
                        RandomForestClassifier())
pipe_rf.fit(X_train, y_train)
print("training score: ", round(pipe_rf.score(X_train, y_train), 3))
print("validation score: ", round(pipe_rf.score(X_dev, y_dev), 3))

### Prediction on dev set

X_dev_hotel_type_pred = pipe_svc.predict(X_dev)

dev_hotel_type_pred_df = pd.DataFrame(
    {
        "utts": X_dev,
        "hotel-type": X_dev_hotel_type_pred,
    }
)

empty_dev = dev_df[["utts"]]

combined_dev = empty_dev.join(dev_hotel_type_pred_df, rsuffix="_hotel").replace(
    "dontknow", np.nan
)

combined_dev.pop('utts_hotel')

combined_dev.to_csv('dev_hotel_type_pred.csv')

### Prediction on TEST set

test_df = read_df_with_pred_intent('dioData_test.csv', 'test_binary_intent_pred.txt')

hotel_test_df = test_df[test_df['intent'] == 'hotel']

X_test = hotel_test_df['utts']

import warnings
warnings.filterwarnings("ignore")

hotel_test_df['hotel-type'] = pipe_svc.predict(X_test)

combined_dev = test_df[['utts']].join(hotel_test_df, rsuffix="_hotel").replace(
    "dontknow", np.nan
)

combined_dev[
    ["utts", "hotel-type"]
].to_csv("test_hotel_type_pred.csv")

## 2. Hotel-pricerange/Restaraunt-pricerange 

train_utt = train_df["utts"].tolist()
dev_utt = dev_df["utts"].tolist()
test_utt = test_df["utts"].tolist()

### 2.1. Restaraunt-pricerange 

rest_preds = train_df["restaurant-pricerange"].tolist()
dev_rest_preds = dev_df["restaurant-pricerange"].tolist()

### SVC

pipe_SVC_rest = make_pipeline(CountVectorizer(), SVC())
pipe_SVC_rest.fit(train_utt, rest_preds)
print("training score: ", pipe_SVC_rest.score(train_utt, rest_preds))
print("validation score: ", pipe_SVC_rest.score(dev_utt, dev_rest_preds))

### Logistic Regression

from sklearn.linear_model import LogisticRegression
pipe_lr_rest = make_pipeline(CountVectorizer(ngram_range = (1,2), 
                                        lowercase = True,
                                        stop_words = 'english'),
                        LogisticRegression())
pipe_lr_rest.fit(train_utt, rest_preds)
print("training score: ", pipe_lr_rest.score(train_utt, rest_preds))
print("validation score: ", pipe_lr_rest.score(dev_utt, dev_rest_preds))

### RandomForest (best)

from sklearn.ensemble import RandomForestClassifier
pipe_rf_rest = make_pipeline(CountVectorizer(ngram_range = (1, 2), 
                                        lowercase = True, 
                                        stop_words = 'english'),
                                        # max_features = 1000), 
                        RandomForestClassifier())
pipe_rf_rest.fit(train_utt, rest_preds)
print("training score: ", pipe_rf_rest.score(train_utt, rest_preds))
print("validation score: ", pipe_rf_rest.score(dev_utt, dev_rest_preds))

### 2.2. Hotel-pricerange

hotel_preds = train_df["hotel-pricerange"].tolist()
dev_hotel_preds = dev_df["hotel-pricerange"].tolist()

pipe = make_pipeline(CountVectorizer(), DummyClassifier())
pipe.fit(train_utt, hotel_preds)
pipe.score(train_utt, hotel_preds)

### SVC

pipe_SVC_hotel = make_pipeline(CountVectorizer(), SVC())
pipe_SVC_hotel.fit(train_utt, hotel_preds)
print("training score: ", pipe_SVC_hotel.score(train_utt, hotel_preds))
print("validation score: ", pipe_SVC_hotel.score(dev_utt, dev_hotel_preds))

### Logistic Regression (best)

pipe_lr_hotel = make_pipeline(CountVectorizer(ngram_range = (1,2), 
                                        lowercase = True,
                                        stop_words = 'english'),
                        LogisticRegression())
pipe_lr_hotel.fit(train_utt, hotel_preds)
print("training score: ", pipe_lr_hotel.score(train_utt, hotel_preds))
print("validation score: ", pipe_lr_hotel.score(dev_utt, dev_hotel_preds))

### RandomForest

pipe_rf_hotel = make_pipeline(CountVectorizer(ngram_range = (1, 2), 
                                        lowercase = True, 
                                        stop_words = 'english'),
                                        #max_features = 1000), 
                        RandomForestClassifier())
pipe_rf_hotel.fit(train_utt, hotel_preds)
print("training score: ", pipe_rf_hotel.score(train_utt, hotel_preds))
print("validation score: ", pipe_rf_hotel.score(dev_utt, dev_hotel_preds))

## 3. Combine and output hotel/restaraunce-pricerange

final = pd.DataFrame({"utts":dev_utt,
              "hotel-pricerange": pipe_lr_hotel.predict(dev_utt).tolist(),
              "restaurant-pricerange": pipe_rf_rest.predict(dev_utt).tolist()})
final.to_csv('dev_price_pred.csv')

final_test = pd.DataFrame({"utts":test_utt,
              "hotel-pricerange": pipe_lr_hotel.predict(test_utt).tolist(),
              "restaurant-pricerange": pipe_rf_rest.predict(test_utt).tolist()})
final_test.to_csv('test_price_pred.csv')
